requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
PyMuPDF==1.23.26
python-docx==1.1.0
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
pymongo==4.6.3
APScheduler==3.10.4
cryptography==42.0.8
psutil==5.9.8
pytest==8.2.0
pytest-cov==5.0.0
python-dotenv==1.0.1